% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_total_var.R
\name{bayes_total_var}
\alias{bayes_total_var}
\title{Calculate the variance of gamma using the law of total variance}
\usage{
bayes_total_var(
  V_x,
  V_y,
  mu_b,
  mu2_b,
  alpha_b,
  mu_a,
  mu2_a,
  alpha_a,
  Gy_t_Gy,
  Gy_t_y,
  num_samples
)
}
\description{
Calculate the variance of gamma using the law of total variance
}
\keyword{internal}
