% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_total_var_rss.R
\name{bayes_total_var_rss}
\alias{bayes_total_var_rss}
\title{Calculate the variance of gamma using the law of total variance for the
Z-score derivation}
\usage{
bayes_total_var_rss(
  V_x,
  V_y,
  mu_b,
  mu2_b,
  alpha_b,
  mu_a,
  mu2_a,
  alpha_a,
  R,
  Z_y,
  num_samples,
  sigma2_gamma_prior
)
}
\description{
Calculate the variance of gamma using the law of total variance for the
Z-score derivation
}
\keyword{internal}
