% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmr_init.R
\name{fmr_init}
\alias{fmr_init}
\title{Initialize the posteriors for b and alpha}
\usage{
fmr_init(
  M,
  L_x,
  L_y,
  susie_init,
  susie_init_y,
  scaled_prior_variance_x,
  scaled_prior_variance_y,
  varX,
  varY,
  Gx_t_Gx,
  Gy_t_Gy,
  mu_gamma_init,
  mu2_gamma_init
)
}
\description{
Initialize the posteriors for b and alpha
}
\keyword{internal}
