% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs_rss.R
\name{get_cs_rss}
\alias{get_cs_rss}
\title{Get credible sets for exposure or outcome for the Z-score implementation}
\usage{
get_cs_rss(V, alpha, R)
}
\arguments{
\item{V}{The L x M matrix of prior variance estimates}

\item{alpha}{The list of length M of L x p matrices of pips for each region}

\item{R}{The list of length M of correlation matrices}
}
\value{
A list of lists containing the credible sets for each region
}
\description{
Function that gets credible sets for the exposure or outcome.
Note that much of the code is directly copied from the susieR package, but
modified so that a SuSiE object is not needed
}
