% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbo_ss.R
\name{get_ER2_ss}
\alias{get_ER2_ss}
\title{Get the ELBO when there is only sufficient statistics}
\usage{
get_ER2_ss(XtX, Xty, alpha, mu, mu2)
}
\description{
Get the ELBO when there is only sufficient statistics
}
\keyword{internal}
