% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_methods.R
\name{negloglik_rss_alpha}
\alias{negloglik_rss_alpha}
\title{Relevant functions for running finemappingMR with Z-score derivation}
\usage{
negloglik_rss_alpha(sl2, RRinvR, Z_star_l, n_y)
}
\description{
Relevant functions for running finemappingMR with Z-score derivation
}
\keyword{internal}
