% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleOverlap_methods.R
\name{negloglik_sampleOverlap_alpha}
\alias{negloglik_sampleOverlap_alpha}
\title{Relevant functions for running finemappingMR with sample overlap}
\usage{
negloglik_sampleOverlap_alpha(sl2, RRinvR, Z_star_l, n_y, rho)
}
\description{
Relevant functions for running finemappingMR with sample overlap
}
\keyword{internal}
