% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_freq_method_ss_dev.R
\name{run_freq_method_ss_dev}
\alias{run_freq_method_ss_dev}
\title{Frequentist MR estimate using sufficient statistics for exposure and outcome.}
\usage{
run_freq_method_ss_dev(
  Gx_t_Gx,
  Gx_t_x,
  xtx,
  Gy_t_Gy,
  Gy_t_y,
  yty,
  n_x,
  n_y,
  L_x = 10,
  L_y = 10,
  scaled_prior_variance_x = 0.2,
  scaled_prior_variance_y = 0.2,
  estimate_prior_variance_x = TRUE,
  estimate_prior_variance_y = TRUE,
  residual_variance_x = NULL,
  residual_variance_y = NULL,
  estimate_residual_variance_x = FALSE,
  estimate_residual_variance_y = FALSE,
  tol = 1e-04,
  max_iter = 1000,
  calc_cs_x = FALSE,
  calc_cs_y = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Gx_t_Gx}{A list of matrices \eqn{G_x'G_x} in which the columns of \eqn{G_x}
are centered to have mean zero. Exposure data.}

\item{Gx_t_x}{A list of matrices \eqn{G_x'x} in which the columns of \eqn{G_x}
are centered to have mean zero. Exposure data.}

\item{xtx}{A scalar \eqn{x'x} in which x is centered to have mean 0. Exposure data.}

\item{Gy_t_Gy}{A list of matrices \eqn{G_y'G_y} in which the columns of \eqn{G_y}
are centered to have mean zero. Outcome data.}

\item{Gy_t_y}{A list of matrices \eqn{G_y'y} in which the columns of \eqn{G_y}
are centered to have mean zero. Outcome data.}

\item{yty}{A scalar \eqn{y'y} in which y is centered to have mean 0. Outcome data.}

\item{n_x}{The exposure sample size.}

\item{n_y}{The outcome sample size.}

\item{L_x}{Maximum number of non-zero effects for the exposure.}

\item{L_y}{Maximum number of non-zero effects for the outcome.}

\item{scaled_prior_variance_x}{The prior variance, divided by
\code{var(x)}; that is, the prior variance of each
non-zero element of b is \code{var(x) * scaled_prior_variance_x}. The
value provided should be either a scalar or a vector of length
\code{L_x}. If \code{estimate_prior_variance_x = TRUE}, this provides
initial estimates of the prior variances.}

\item{scaled_prior_variance_y}{The prior variance, divided by
\code{var(y)}; that is, the prior variance of each
non-zero element of alpha is \code{var(y) * scaled_prior_variance_y}. The
value provided should be either a scalar or a vector of length
\code{L_x}. If \code{estimate_prior_variance_y = TRUE}, this provides
initial estimates of the prior variances.}

\item{estimate_prior_variance_x}{If \code{estimate_prior_variance_x =
  TRUE}, the prior variance is estimated (this is a separate
parameter for each of the L_x effects). If provided,
\code{scaled_prior_variance_x} is then used as an initial value for
the optimization which is done using \code{optim}.
When \code{estimate_prior_variance_x = FALSE}, the
prior variance for each of the L_x effects is determined by the
value supplied to \code{scaled_prior_variance_x}.}

\item{estimate_prior_variance_y}{If \code{estimate_prior_variance_y =
  TRUE}, the prior variance is estimated (this is a separate
parameter for each of the L_y effects). If provided,
\code{scaled_prior_variance_y} is then used as an initial value for
the optimization which is done using \code{optim}.
When \code{estimate_prior_variance_y = FALSE}, the
prior variance for each of the L_y effects is determined by the
value supplied to \code{scaled_prior_variance_y}. Currently has
no functionality.}

\item{residual_variance_x}{The variance of the exposure residual. If \code{estimate_residual_variance_x = TRUE},
this is used as an initial value.}

\item{residual_variance_y}{The variance of the outcome residual. If \code{estimate_residual_variance_y = TRUE},
this is used as an initial value.}

\item{estimate_residual_variance_x}{If \code{TRUE}, exposure residual variance is estimated
using initial value \code{residual_variance_x}. Currently has
no functionality.}

\item{estimate_residual_variance_y}{If \code{TRUE}, outcome residual variance is estimated
using initial value \code{residual_variance_y}. Currently has
no functionality.}

\item{tol}{A small, non-negative number specifying the convergence
tolerance for the variational Bayes procedure. The fitting procedure
will halt when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is
less than \code{tol}.}

\item{max_iter}{Maximum number of iterations before stopping estimation procedure.}

\item{calc_cs_x}{Calculate credible sets for the exposure? Default = FALSE.}

\item{calc_cs_y}{Calculate credible sets for the outcome? Default = FALSE.}

\item{verbose}{Output progress? Default = FALSE.}
}
\value{
A list containing various results from the estimation procedure,
including a data frame containing the gamma estimation results, as well as
the posterior first and second moments for b and alpha, the prior non-zero
effect variance estimates, and the PIPs for each non-zero effect, and the
vector of ELBO iterations.
}
\description{
Function that performs our frequentist MR estimation method using
sufficient statistics for exposure and outcome. This is a development version
of run_freq_method_ss.R, where I am trying to improve efficiency.
}
