% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_posterior_rss.R
\name{update_posterior_rss}
\alias{update_posterior_rss}
\title{Update the posteriors for b and alpha using Z-score derivation}
\usage{
update_posterior_rss(Z_star_l, omega_j)
}
\description{
Update the posteriors for b and alpha using Z-score derivation
}
\keyword{internal}
